Fix values for real numbers when given by SmtInterpol. We truncated them to integers.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@13603 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/predicates/smtInterpol/SmtInterpolNumeralFormulaManager.java
@@ -83,7 +83,7 @@ abstract class SmtInterpolNumeralFormulaManager
     Term arg = SmtInterpolUtil.getArg(pNumber, 0);
     if (SmtInterpolUtil.isNumber(arg)) {
       // TODO: BUG: possible bug
-      return SmtInterpolUtil.toNumber(arg) == -1;
+      return SmtInterpolUtil.toNumber(arg).doubleValue() == -1;
     }
     return false;
   }
